--- geomesa-security/src/test/java/org/locationtech/geomesa/security/TestAuthorizationsProvider.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2015 Commonwealth Computer Research, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.locationtech.geomesa.security;
-
-import org.apache.accumulo.core.security.Authorizations;
-
-import java.io.Serializable;
-import java.util.Map;
-
-/**
- * Test authorizations provider that doesn't use any auths
- */
-public class TestAuthorizationsProvider
-        implements AuthorizationsProvider {
-    @Override
-    public Authorizations getAuthorizations() {
-        return null;
-    }
-
-    @Override
-    public void configure(Map<String, Serializable> params) {
-
-    }
-}
