+++ geomesa-security/src/test/resources/META-INF/services/org.locationtech.geomesa.security.AuthorizationsProvider
@@ -1,0 +1,1 @@
+org.locationtech.geomesa.security.TestAuthorizationsProvider
