+++ pom.xml
@@ -913,7 +913,7 @@
             <dependency>
                 <groupId>org.locationtech.geomesa</groupId>
                 <artifactId>geomesa-web-security-accumulo1.5</artifactId>
-                <version>1.0.0-rc.6-SNAPSHOT</version>
+                <version>${project.version}</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
